src/roberts_blog/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-profanity-filter==0.2.1
gunicorn==20.1.0
inflection==0.5.1
Pillow==9.3.0
python-decouple==3.6
sqlparse==0.4.3
whitenoise==6.2.0
